+++ OneBlink.SDK/models/FormElement.cs
@@ -102,6 +102,8 @@ namespace OneBlink.SDK.Model
         public List<Guid> elementIds { get; set; }
 
         public string placeholderValue { get; set; }
+        public int? minLength {get; set; }
+        public int? maxLength {get; set; }
 
         public Boolean? isInteger {get; set; }
         private static Guid initialiseId(Guid? id)
@@ -126,7 +128,9 @@ namespace OneBlink.SDK.Model
             bool required = false,
             bool readOnly = false,
             string defaultValue = null,
-            string placeholderValue = null
+            string placeholderValue = null,
+            int? minLength = null,
+            int? maxLength = null
         )
         {
             FormElement textElement = new FormElement();
@@ -141,6 +145,8 @@ namespace OneBlink.SDK.Model
             textElement.readOnly = readOnly;
             textElement.defaultValue = defaultValue;
             textElement.placeholderValue = placeholderValue;
+            textElement.minLength = minLength;
+            textElement.maxLength = maxLength;
             return textElement;
         }